test: fix merge
