+++ packages/client/test/test-utils/utils.ts
@@ -25,8 +25,6 @@ import { StreamrClientConfig } from '../../src/Config'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { EncryptionUtil } from '../../src/encryption/EncryptionUtil'
 import { addAfterFn } from './jest-utils'
-import { StreamDefinition } from '../../src/types'
-import { PublishMetadata } from '../../src/publish/PublishPipeline'
 
 const testDebugRoot = Debug('test')
 const testDebug = testDebugRoot.extend.bind(testDebugRoot)
@@ -183,17 +181,4 @@ export const createMockMessage = (
     }
     msg.signature = SigningUtil.sign(msg.getPayloadToSign(StreamMessage.SIGNATURE_TYPES.ETH), opts.publisher.privateKey)
     return msg
-}
-
-export async function* publishFromMetadata<T>(
-    streamDefinition: StreamDefinition, 
-    seq: AsyncIterable<PublishMetadata<T>>,
-    client: StreamrClient
-): AsyncGenerator<StreamMessage<T>, void, unknown> {
-    for await (const msg of seq) {
-        yield await client.publish(streamDefinition, msg.content, {
-            timestamp: msg.timestamp,
-            partitionKey: msg.partitionKey
-        })
-    }
 }
